# Class name resolver performance optimisation