<!-- Fill in the relevant information below to help triage your pull request. -->

|      Q       |   A
|------------- | -----------
| Type         | improvement
| BC Break     | yes
| Fixed issues | 

#### Summary

cc @watari

#1908 suggests a number of changes to improve document performance. I was especially curious about the changes to the document manager, where an additional cache for `ClassMetadata` is added. Looking at that change in isolation, it turns out that most of the improvements to the performance test suite come from that single change.

Upon further investigation, I was able to pinpoint the performance regression to the class name resolution for proxy objects, which now was added to MongoDB ODM until the doctrine/persistence layer supports multiple proxy implementations.

This PR removes this issue by adding a caching version of a class name resolver which removes the need for repeatedly resolving the same class names. Since the original resolver was not marked as `@internal`, it is only deprecated and not removed entirely. The new implementations are marked as `@internal` as users shouldn't need to instantiate those directly. The associated interface is not marked as `@internal` and is covered by the BC promise.

I'll take a look at the other changes in #1908 in separate PRs or merge it directly with the DocumentManager change removed.

For reference, the following times are from before and after the change:
```
⅀T: 9,645.960μs μSD/r 0.000μs μRSD/r: 0.000%
⅀T: 8,144.470μs μSD/r 0.000μs μRSD/r: 0.000%
```

This is an improvement of ~15.5% just for this change alone!